Get rid of Makefiles, use make.py, update USB stack.
Having two make systems (Makefiles AND make.py) just leads to a lot of
friction, so let's fall back to use make.py since it also works on
Windows and we don't really gain anything from Make anyways.

While we're at it, we've also merged the DFU stub endpoint into the
tinydfu-bootloader project, so we should be able to reuse the same
USB stack for both bootloader and the user bitstream.