Use separate VID/PID for bootloader.
When switching between DFU and runtime modes, the badge needs a
different VID/PID for Windows do be able to enumerate its drivers
correctly. We solve this by adding some macros to the yosys command
line.

Note, that we still need to get real VID/PID numbers allocated for
both the badge and its bootloader.